✨ Slide 이동 기능 보완
- 슬라이드 4개씩 이동으로 수정
- cardsInfo에서 slideMockData로 변경
