+++ src/components/MenuSection/slideMockData.js
@@ -1,0 +1,74 @@
+export default [
+  {
+    size: 'MEDIUM',
+    imageURL: 'http://public.codesquad.kr/jk/storeapp/data/main/1155_ZIP_P_0081_T.jpg',
+    title: '오리 주물럭_반조리1',
+    desc: '감질맛 나는 매콤한 양념',
+    curPrice: '12,640원',
+    prevPrice: '15,800원',
+    tags: ['이벤트특가', '런칭특가']
+  },
+  {
+    size: 'MEDIUM',
+    imageURL: 'http://public.codesquad.kr/jk/storeapp/data/main/1155_ZIP_P_0081_T.jpg',
+    title: '오리 주물럭_반조리2',
+    desc: '감질맛 나는 매콤한 양념',
+    curPrice: '12,640원',
+    prevPrice: '15,800원',
+    tags: ['이벤트특가', '런칭특가']
+  },
+  {
+    size: 'MEDIUM',
+    imageURL: 'http://public.codesquad.kr/jk/storeapp/data/main/1155_ZIP_P_0081_T.jpg',
+    title: '오리 주물럭_반조리3',
+    desc: '감질맛 나는 매콤한 양념',
+    curPrice: '12,640원',
+    prevPrice: '15,800원',
+    tags: ['이벤트특가', '런칭특가']
+  },
+  {
+    size: 'MEDIUM',
+    imageURL: 'http://public.codesquad.kr/jk/storeapp/data/main/1155_ZIP_P_0081_T.jpg',
+    title: '오리 주물럭_반조리4',
+    desc: '감질맛 나는 매콤한 양념',
+    curPrice: '12,640원',
+    prevPrice: '15,800원',
+    tags: ['이벤트특가', '런칭특가']
+  },
+  {
+    size: 'MEDIUM',
+    imageURL: 'http://public.codesquad.kr/jk/storeapp/data/main/1155_ZIP_P_0081_T.jpg',
+    title: '오리 주물럭_반조리5',
+    desc: '감질맛 나는 매콤한 양념',
+    curPrice: '12,640원',
+    prevPrice: '15,800원',
+    tags: ['이벤트특가', '런칭특가']
+  },
+  {
+    size: 'MEDIUM',
+    imageURL: 'http://public.codesquad.kr/jk/storeapp/data/main/1155_ZIP_P_0081_T.jpg',
+    title: '오리 주물럭_반조리6',
+    desc: '감질맛 나는 매콤한 양념',
+    curPrice: '12,640원',
+    prevPrice: '15,800원',
+    tags: ['이벤트특가', '런칭특가']
+  },
+  {
+    size: 'MEDIUM',
+    imageURL: 'http://public.codesquad.kr/jk/storeapp/data/main/1155_ZIP_P_0081_T.jpg',
+    title: '오리 주물럭_반조리7',
+    desc: '감질맛 나는 매콤한 양념',
+    curPrice: '12,640원',
+    prevPrice: '15,800원',
+    tags: ['이벤트특가', '런칭특가']
+  },
+  {
+    size: 'MEDIUM',
+    imageURL: 'http://public.codesquad.kr/jk/storeapp/data/main/1155_ZIP_P_0081_T.jpg',
+    title: '오리 주물럭_반조리8',
+    desc: '감질맛 나는 매콤한 양념',
+    curPrice: '12,640원',
+    prevPrice: '15,800원',
+    tags: ['이벤트특가', '런칭특가']
+  }
+];
